hackothon/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
joblib==1.3.2
python-multipart==0.0.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
google-generativeai>=0.8.0
googlemaps>=4.10.0
twilio>=9.0.0
fpdf2>=2.7.0
tensorflow>=2.15.0
xgboost>=2.0.0
catboost>=1.2.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python
scikit-image
scipy
